sili_telegram_bot: Move mediawiki_api to models
  Better fit, i think.
